+++ TgDownloaderCore/GlobalUsings.cs
@@ -1,0 +1,8 @@
+﻿// This is an independent project of an individual developer. Dear PVS-Studio, please check it.
+// PVS-Studio Static Code Analyzer for C, C++, C#, and Java: http://www.viva64.com
+
+global using System.Diagnostics;
+global using System.IO.Enumeration;
+global using System.Reflection;
+global using System.Text;
+global using System.Text.RegularExpressions;